pyproject.toml
[project]
name = "speech-to-latex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.30.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.2",
]
